Use unsafeScientificFromNonNormalized to efficiently construct Scientifics
The scientific package switched to normalizing Scientific numbers on
construction. This means constructing a Scientific using
Data.Scientific.scientific is less efficient than it used to be
because it has to perform normalization.

Fortunately we can recover the loss in performance by using
Data.Scientific.Unsafe.unsafeScientificFromNonNormalized which accepts
a possible non-normalized coefficient and the number of trailing 0s in
the coefficient. This patch modifies the scientific parsers such that
they count the number of trailing 0s.